Fix warning on 64 bit platforms introduced by badnull.cocci.
Clean up ridiculous amount of casting and parentheses in this statement.

git-svn-id: http://svn.enlightenment.org/svn/e/trunk@51486 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ E16/e/epp/cpplib.c
@@ -4023,8 +4023,7 @@ do_pragma(cpp_reader * pfile, struct directive *keyword, unsigned char *buf,
 
 	fname = p + 1;
 	p = strchr(fname, '\"');
-	fname_len =
-	   (int)(((int)(p)) ? ((int)(p - fname)) : ((int)(strlen(fname))));
+	fname_len = (p) ? (int)(p - fname) : (int)strlen(fname);
 
 	for (ptr = pfile->all_include_files; ptr; ptr = ptr->next)
 	  {
